# _Tools/nswag-converter/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "OmegaExplorer API",
    "description": "API pour OmegaExplorer",
    "version": "v1"
  },
  "paths": {
    "/api/admin/generate/map": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "GenerateMap",
        "parameters": [
          {
            "name": "clean",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Put **_ONLY_** your JWT Bearer token on textbox below!",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}